Add 'usable' property/accessor to MemoryStats
Replicating the second line of `free -m` adding buffers + cache to the
free in a seperate property/accessor as to not mess with the current
behaviour of getFreeBytes(). Makes the 'free' much more useful because
it isn't a race to zero bytes (the OS will always try to pack memory
with cache which makes just plain free a not very useful number).