Fix process command and name being incorrect on Windows.
While iterating over the process list, any process for which we didn't have
permission to examine their module name would get a module name from the
last process that we did have permission to access.

This change fixes the problem, all processes that we can't get the name
of will now be reported as "<unknown>".

A similar fix for the process command line was also added.